// apps/web/src/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[];

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.4";
  };
  public: {
    Tables: {
// ...
      waitlist_signups: {
        Row: {
          campaign: string;
          confirmed_at: string | null;
          created_at: string;
          email: string;
          full_name: string | null;
          id: string;
          ip: unknown;
          metadata: Json;
          source: string | null;
          status: Database["public"]["Enums"]["waitlist_status"];
          user_agent: string | null;
        };
        Insert: {
          campaign?: string;
          confirmed_at?: string | null;
          created_at?: string;
          email: string;
          full_name?: string | null;
          id?: string;
          ip?: unknown;
          metadata?: Json;
          source?: string | null;
          status?: Database["public"]["Enums"]["waitlist_status"];
          user_agent?: string | null;
        };
        Update: {
          campaign?: string;
          confirmed_at?: string | null;
          created_at?: string;
          email?: string;
          full_name?: string | null;
          id?: string;
          ip?: unknown;
          metadata?: Json;
          source?: string | null;
          status?: Database["public"]["Enums"]["waitlist_status"];
          user_agent?: string | null;
        };
        Relationships: [];
      };
    };
// ...
  };
};

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">;

type DefaultSchema = DatabaseWithoutInternals[Extract<
  keyof Database,
  "public"
>];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never;